Release v2xx

+++ embedded-software/mcu-primary/src/application/sox/soc_lstm/socEstimator_wrapper.h
@@ -1,6 +1,6 @@
 #ifndef SOC_ESTIMATOR_WRAPPER_H
 #define SOC_ESTIMATOR_WRAPPER_H
 
-void SOX_GetStateOfCharge_fromLSTM(float *voltage, float *current, float *temperature, int len, float *soc_out);
+void SOX_GetStateOfCharge_fromLSTM(double *voltage, double *current, double *temperature, int len, double *soc_out);
 
 #endif
